Refactor ProfileEditModal to use SlideModal component
- Replace custom animated modal with reusable SlideModal
- Simplify modal rendering and animation logic
- Remove redundant state and animation management
